move mediaQueries from components to helpers folder
